Set sync_dist=True when logging best so far validation accuracy
We set sync_dist=True for logging best so far validation accuracy since this metric is
logged through value (instead of metric object like all the other metrics in template),
so this makes sure it will be synchronized accross devices in multi-gpu setup.

As far as I'm aware, this change doesn't affect the result value of computed metric in
this specific case, because synchronization doesn't matter when logging best so far value
which is retrieved from other metric object which is already synchronized.

But it's a good practice to indicate this is what should be done when logging trough value.